requirements.txt
fastapi[all]==0.119.0
openai>=1.0.0
uvicorn==0.37.0
sqlalchemy==2.0.44
asyncpg==0.30.0
pytest==8.4.2
pytest-asyncio==1.2.0
faiss-cpu==1.12.0
requests==2.32.5
pgvector==0.4.1
psycopg==3.2.10
aiofiles==25.1.0
docling>=2.70.0
PyYAML==6.0.2
greenlet==3.0.3
